Force redraw in appear handler
Explicitly invokes handle_tick() from the appear handler so that the window
is redrawn quickly after the appear handler runs.  This was necessary in
order to get rid of a delay seen in between backing out of the watch menu
and the date string reappearing on the watch face.

This commit does away with a redundant assignment to the global struct tm
*now, which shouldn't be necessary, but which could cause problems if my
assumption is incorrect that the appear handler will always be invoked
before any layer update procedures are run.